provisioning/test/requirements.txt
pytest
google-cloud-run
httpx
google-cloud-secret-manager
pytest-playwright